Refactor code to deal with circular dependency
